feat: M4+M5 事件体系与 UI 层全量迁移 C#（回归 55/55）
M4 事件体系：GameEvent/FogEvent 继承链 + GameEventManager/FogEventManager
+ EliteTurretEvent/FormationStrikeEvent/Mothership 等 21 个 .cs，删 17 个 .gd
M5 UI 层：UITheme/ChamferedPanel/SegmentedBar/BuffIcons 基类 + 14 场景脚本
+ TestExit 测试退出辅助，删 19 个 .gd，4 场景切 C#
修复：Tutorial AddChild 缺失、退出 segfault（GC 时机→TestExit.Quit）、
维修浮点 flake（BaseConsole/PlayerDamage 全程 double）、const preload 重载解析

+++ csharp/godot/PlayerDamage.cs
@@ -110,7 +110,9 @@ public class PlayerDamage
         }
         else if (SinceDamage >= GameStateBridge.Call("passive_regen_delay").AsDouble())
         {
-            GameStateBridge.Call("heal", (float)GameStateBridge.Call("passive_regen_rate").AsDouble() * delta);
+            // M5：(float) 截断致 health 被 float 精度污染（50+9.2→59.2000004798174），
+            // 后续 base 维修 heal 差值不精确回满（smoke flake 根因）；全程 double
+            GameStateBridge.Call("heal", GameStateBridge.Call("passive_regen_rate").AsDouble() * delta);
         }
     }
